Refactor header tabs into HeaderTab component
Will be useful for next change that introduces stylish underlines under the active tab
